fix: move resume S3 upload to service layer and prevent ATS check from overwriting master profile
